+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Changed
 * Exclude private constructors by default. [#70](https://github.com/nojaf/telplin/issues/70)
 
+### Fixed
+* Private constructor in struct. [#79](https://github.com/nojaf/telplin/issues/79)
+
 ## 0.7.0 - 05-06-23
 
 ### Changed